+++ packages/pds/src/db/database-schema.ts
@@ -26,7 +26,7 @@ import * as messageQueue from './tables/message-queue'
 import * as messageQueueCursor from './tables/message-queue-cursor'
 import * as moderation from './tables/moderation'
 import * as mute from './tables/mute'
-import * as sequencedEvents from './tables/sequenced-event'
+import * as repoSeq from './tables/repo-seq'
 
 export type DatabaseSchemaType = user.PartialDB &
   didHandle.PartialDB &
@@ -57,7 +57,7 @@ export type DatabaseSchemaType = user.PartialDB &
   messageQueueCursor.PartialDB &
   moderation.PartialDB &
   mute.PartialDB &
-  sequencedEvents.PartialDB
+  repoSeq.PartialDB
 
 export type DatabaseSchema = Kysely<DatabaseSchemaType>
