+++ packages/pds/src/db/tables/repo-seq.ts
@@ -1,0 +1,11 @@
+import { Generated } from 'kysely'
+
+export interface RepoSeq {
+  seq: Generated<number>
+  did: string
+  commit: string
+}
+
+export const tableName = 'repo_seq'
+
+export type PartialDB = { [tableName]: RepoSeq }
