+++ packages/pds/src/index.ts
@@ -26,6 +26,7 @@ import { BlobDiskCache, ImageProcessingServer } from './image/server'
 import { createServices } from './services'
 import { createHttpTerminator, HttpTerminator } from 'http-terminator'
 import AppContext from './context'
+import Sequencer from './sequencer'
 
 export type { ServerConfigValues } from './config'
 export { ServerConfig } from './config'
@@ -59,6 +60,7 @@ export class PDS {
     })
 
     const messageQueue = new SqlMessageQueue('pds', db)
+    const sequencer = new Sequencer(db)
 
     const mailTransport =
       config.emailSmtpUrl !== undefined
@@ -104,6 +106,7 @@ export class PDS {
       cfg: config,
       auth,
       messageQueue,
+      sequencer,
       services,
       mailer,
       imgUriBuilder,
